Struct pwasm_std::types::H256

#[repr(C)]

pub struct H256(_);

A 256-bits (32 bytes) hash type.

Methods

impl H256

pub fn repeat_byte(byte: u8) -> H256

Returns a new fixed hash where all bits are set to the given byte.

pub fn zero() -> H256

Returns a new zero-initialized fixed hash.

pub fn len_bytes() -> usize

Returns the size of this hash in bytes.

pub fn as_bytes(&self) -> &[u8]

Extracts a byte slice containing the entire fixed hash.

pub fn as_bytes_mut(&mut self) -> &mut [u8]

Extracts a mutable byte slice containing the entire fixed hash.

pub fn as_fixed_bytes(&self) -> &[u8; 32]

Extracts a reference to the byte array containing the entire fixed hash.

pub fn as_fixed_bytes_mut(&mut self) -> &mut [u8; 32]

Extracts a reference to the byte array containing the entire fixed hash.

pub fn to_fixed_bytes(self) -> [u8; 32]

Returns the inner bytes array.

pub fn as_ptr(&self) -> *const u8

Returns a constant raw pointer to the value.

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns a mutable raw pointer to the value.

pub fn assign_from_slice(&mut self, src: &[u8])

Assign the bytes from the byte slice src to self.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in self do not match.

pub fn from_slice(src: &[u8]) -> Self

Create a new fixed-hash from the given slice src.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn covers(&self, b: &Self) -> bool

Returns true if all bits set in b are also set in self.

pub fn is_zero(&self) -> bool

Returns true if no bits are set.